Make an internal transaction class
